+++ src/EasyModelServiceProvider.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Ramadan\EasyModel;
+
+use Illuminate\Support\ServiceProvider;
+
+class EasyModelServiceProvider extends ServiceProvider
+{
+    /**
+     * Register any application services.
+     *
+     * @return void
+     */
+    public function register()
+    {
+        $this->app->singleton(EasyModel::class, fn() => new EasyModel);
+    }
+}