+++ src/Exceptions/EasyModelException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Ramadan\EasyModel\Exceptions;
+
+/**
+ * Marker interface implemented by every exception thrown by the package.
+ */
+interface EasyModelException
+{
+}