+++ src/Updatable.php
@@ -20,9 +20,9 @@ trait Updatable
     protected $modelForUpdate;
 
     /**
-     * The search or update builder.
+     * The underlying queryable instance.
      *
-     * @var \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder
+     * @return \Illuminate\Database\Eloquent\Model|\Illuminate\Database\Eloquent\Builder|\Illuminate\Database\Query\Builder
      */
     protected $searchOrUpdateQuery;
 
@@ -30,19 +30,23 @@ trait Updatable
      * Get an updatable eloquent builder.
      *
      * @param  string|null  $relationship
-     * @return \Illuminate\Database\Eloquent\Builder
+     * @return \Illuminate\Database\Eloquent\Model|\Illuminate\Database\Eloquent\Builder
      *
      * @throws \Ramadan\EasyModel\Exceptions\InvalidModel
      */
     protected function getUpdatableEloquentBuilder($relationship = null)
     {
-        if (empty($this->getUpdatableModel())) {
-            throw new InvalidModel("You must set the updatable model first.");
+        $model = $this->getUpdatableModel();
+
+        if (empty($model)) {
+            throw InvalidModel::updatableNotSet();
         }
 
-        return empty($relationship) ?
-            $this->getUpdatableModel()->newQuery() :
-            $this->getUpdatableModel()->{$relationship}()->getQuery();
+        if ($model->exists) {
+            return $model;
+        }
+
+        return empty($relationship) ? $model->newQuery() : $model->{$relationship}()->getQuery();
     }
 
     /**
@@ -102,7 +106,7 @@ trait Updatable
 
         // If a model has been created or updated, it takes precedence. In such cases,
         // we will increment the values of its columns.
-        if (!empty($this->modelForUpdate)) {
+        if (! empty($this->modelForUpdate)) {
             foreach ($attributes as $column => $value) {
                 $this->modelForUpdate->{$column} += $value;
             }
@@ -115,9 +119,9 @@ trait Updatable
         /**
          * @see https://php.net/manual/en/closure.call.php
          */
-        $extra = !$usingQueryBuilder ?
-            (fn($args) => $this->addUpdatedAtColumn($args))->call($this->getSearchOrUpdateBuilder(), []) :
-            [];
+        $extra = ! $usingQueryBuilder
+            ? (fn($args) => $this->addUpdatedAtColumn($args))->call($this->getSearchOrUpdateBuilder(), [])
+            : [];
 
         $this->searchOrUpdateQuery->incrementEach($attributes, $extra);
 
@@ -139,7 +143,7 @@ trait Updatable
 
         // If a model has been created or updated, it takes precedence. In such cases,
         // we will decrement the values of its columns.
-        if (!empty($this->modelForUpdate)) {
+        if (! empty($this->modelForUpdate)) {
             foreach ($attributes as $column => $value) {
                 $this->modelForUpdate->{$column} -= $value;
             }
@@ -152,9 +156,9 @@ trait Updatable
         /**
          * @see https://php.net/manual/en/closure.call.php
          */
-        $extra = !$usingQueryBuilder ?
-            (fn($args) => $this->addUpdatedAtColumn($args))->call($this->getSearchOrUpdateBuilder(), []) :
-            [];
+        $extra = ! $usingQueryBuilder
+            ? (fn($args) => $this->addUpdatedAtColumn($args))->call($this->getSearchOrUpdateBuilder(), [])
+            : [];
 
         $this->searchOrUpdateQuery->decrementEach($attributes, $extra);
 
@@ -176,7 +180,7 @@ trait Updatable
 
         // If a model has been created or updated, it takes precedence. In such cases,
         // we will zero out the values of its columns.
-        if (!empty($this->modelForUpdate)) {
+        if (! empty($this->modelForUpdate)) {
             $this->modelForUpdate->update(array_fill_keys($attributes, 0));
 
             return $this;
@@ -202,10 +206,10 @@ trait Updatable
 
         // If a model has been created or updated, it takes precedence. In such cases,
         // we will toggle the values of its columns.
-        if (!empty($this->modelForUpdate)) {
+        if (! empty($this->modelForUpdate)) {
             $columns = $this->modelForUpdate->only($attributes);
 
-            $toggle = array_map(fn($value) => !$value, $columns);
+            $toggle = array_map(fn($value) => ! $value, $columns);
 
             $this->modelForUpdate->update($toggle);
 
@@ -226,7 +230,7 @@ trait Updatable
      *
      * @param  string|null  $relationship
      * @param  bool  $isQueryBuilder
-     * @return \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder
+     * @return \Illuminate\Database\Eloquent\Model|\Illuminate\Database\Eloquent\Builder|\Illuminate\Database\Query\Builder
      *
      * @throws \Ramadan\EasyModel\Exceptions\InvalidModel
      */
@@ -234,7 +238,7 @@ trait Updatable
     {
         // If the "setRelationship" method exists, it means the request is coming
         // from the "Searchable" context since the "Updatable" trait is used there.
-        if (!empty($relationship) && method_exists($this, 'setRelationship')) {
+        if (! empty($relationship) && method_exists($this, 'setRelationship')) {
             $this->setRelationship($relationship);
         }
 
@@ -262,7 +266,7 @@ trait Updatable
         $model = $this->getUpdatableModel();
 
         // If the developer has set a model for update, it takes precedence.
-        if (!empty($model) && $model->exists) {
+        if (! empty($model) && $model->exists) {
             $this->modelForUpdate = $model;
         } elseif (empty($this->searchOrUpdateQuery)) {
             $this->searchOrUpdateQuery = $this->getSearchOrUpdateBuilder(isQueryBuilder: $usingQueryBuilder);
@@ -277,11 +281,22 @@ trait Updatable
      */
     public function setUpdatableQuery(QueryBuilder|EloquentBuilder|null $query = null)
     {
-        if ($query instanceof QueryBuilder) {
+        if ($query instanceof EloquentBuilder || $query instanceof QueryBuilder) {
             $this->searchOrUpdateQuery = $query;
-        } elseif ($query instanceof EloquentBuilder) {
-            $this->searchOrUpdateQuery = $query->getQuery();
         }
+
+        return $this;
+    }
+
+    /**
+     * Reset the internal updatable state.
+     *
+     * @return $this
+     */
+    public function flushUpdatable()
+    {
+        $this->searchOrUpdateQuery = null;
+        $this->modelForUpdate      = null;
 
         return $this;
     }
@@ -290,7 +305,7 @@ trait Updatable
      * Fetch the builder instance.
      *
      * @param  bool  $isQueryBuilder
-     * @return \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder
+     * @return \Illuminate\Database\Eloquent\Model|\Illuminate\Database\Eloquent\Builder|\Illuminate\Database\Query\Builder
      *
      * @throws \Ramadan\EasyModel\Exceptions\InvalidModel
      */